misc.c->env.c split: Extract code
In the ongoing effort to reduce the number of loc in misc[12].c,
(see #209) this commit pulls a number of functions related to
reading, writing, and modifying environment variables from misc1.c
into a new module. This will later be merged in to os/env.c, but in an
effort to keep the diffs relatively sane because of the way the work was
orriginally done, this commit and the next 2 will operate on a new file
which will then be merged in to os/env.c
