+++ src/nvim/environment.h
@@ -1,0 +1,10 @@
+#ifndef NVIM_ENVIRONMENT_H
+#define NVIM_ENVIRONMENT_H
+
+#include "nvim/vim.h"
+
+#ifdef INCLUDE_GENERATED_DECLARATIONS
+# include "environment.h.generated.h"
+#endif
+
+#endif
