+++ src/nvim/ex_getln.c
@@ -25,6 +25,7 @@
 #include "nvim/cursor.h"
 #include "nvim/digraph.h"
 #include "nvim/edit.h"
+#include "nvim/environment.h"
 #include "nvim/eval.h"
 #include "nvim/ex_cmds.h"
 #include "nvim/ex_cmds2.h"
