+++ src/nvim/file_search.c
@@ -51,6 +51,7 @@
 
 #include "nvim/vim.h"
 #include "nvim/ascii.h"
+#include "nvim/environment.h"
 #include "nvim/file_search.h"
 #include "nvim/charset.h"
 #include "nvim/fileio.h"
