+++ src/nvim/hardcopy.c
@@ -23,6 +23,7 @@
 #include "nvim/hardcopy.h"
 #include "nvim/buffer.h"
 #include "nvim/charset.h"
+#include "nvim/environment.h"
 #include "nvim/eval.h"
 #include "nvim/ex_cmds2.h"
 #include "nvim/ex_docmd.h"
@@ -31,7 +32,6 @@
 #include "nvim/memline.h"
 #include "nvim/memory.h"
 #include "nvim/message.h"
-#include "nvim/misc1.h"
 #include "nvim/misc2.h"
 #include "nvim/garray.h"
 #include "nvim/option.h"
