+++ src/nvim/if_cscope.c
@@ -15,12 +15,12 @@
 
 #include "nvim/vim.h"
 #include "nvim/ascii.h"
+#include "nvim/environment.h"
 #include "nvim/if_cscope.h"
 #include "nvim/charset.h"
 #include "nvim/eval.h"
 #include "nvim/fileio.h"
 #include "nvim/message.h"
-#include "nvim/misc1.h"
 #include "nvim/misc2.h"
 #include "nvim/memory.h"
 #include "nvim/os/time.h"
