+++ src/nvim/main.c
@@ -20,6 +20,7 @@
 #include "nvim/buffer.h"
 #include "nvim/charset.h"
 #include "nvim/diff.h"
+#include "nvim/environment.h"
 #include "nvim/eval.h"
 #include "nvim/ex_cmds.h"
 #include "nvim/ex_cmds2.h"
