+++ src/nvim/mark.c
@@ -21,6 +21,7 @@
 #include "nvim/buffer.h"
 #include "nvim/charset.h"
 #include "nvim/diff.h"
+#include "nvim/environment.h"
 #include "nvim/eval.h"
 #include "nvim/ex_cmds.h"
 #include "nvim/fileio.h"
@@ -29,7 +30,6 @@
 #include "nvim/memline.h"
 #include "nvim/memory.h"
 #include "nvim/message.h"
-#include "nvim/misc1.h"
 #include "nvim/misc2.h"
 #include "nvim/normal.h"
 #include "nvim/option.h"
