+++ src/nvim/memory.c
@@ -12,6 +12,7 @@
 #include "nvim/charset.h"
 #include "nvim/diff.h"
 #include "nvim/edit.h"
+#include "nvim/environment.h"
 #include "nvim/eval.h"
 #include "nvim/ex_cmds.h"
 #include "nvim/ex_docmd.h"
