+++ src/nvim/screen.c
@@ -94,6 +94,7 @@
 #include "nvim/vim.h"
 #include "nvim/ascii.h"
 #include "nvim/arabic.h"
+#include "nvim/environment.h"
 #include "nvim/screen.h"
 #include "nvim/buffer.h"
 #include "nvim/charset.h"
